requirements.txt
flask==3.0.0
jupyter==1.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
datasets==2.15.0
transformers==4.35.2
matplotlib==3.8.2
seaborn==0.13.0
pymongo==4.6.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0